fix(06.1): constrain TBA rank to an integer at the fetch boundary
WR-01: tbaEventRankingSchema.rank lacked .int(), unlike every other
boundary this value crosses (corpus schema.sql INTEGER column,
pageArtifacts.ts's rank: z.number().int()...). A drifted non-integer
rank would have been silently persisted into a SQLite INTEGER column
instead of throwing at the fetch boundary (T-06.1-01). matches_played
and dq were left unchanged — they are read nowhere else in the
codebase and cross no other integer-typed boundary, so there was no
comparable evidence to extend the constraint to them.

+++ packages/ingest/schemas.ts
@@ -118,7 +118,12 @@ export const tbaMediaListSchema = z.array(tbaMediaSchema);
  * used; see `rankings.ts`).
  */
 export const tbaEventRankingSchema = z.object({
-  rank: z.number(),
+  // `.int()`: every other boundary this value crosses treats it as strictly
+  // integral (`packages/corpus/schema.sql`'s `rank INTEGER NOT NULL`,
+  // `packages/harness/pageArtifacts.ts`'s `rank: z.number().int()...`) — a
+  // non-integral rank must throw here, at the fetch boundary, not silently
+  // persist into a SQLite INTEGER column (T-06.1-01).
+  rank: z.number().int(),
   team_key: z.string(),
   matches_played: z.number(),
   dq: z.number(),
